Create README.md
How do I interpret user needs and implement them into a program? How does creating “user stories” help with this?  

Interpreting user needs and implementing them into a program involves understanding the requirements and expectations of the users of the program, and designing and building the program to meet those needs. One way to do this is by creating "user stories," which are short, simple descriptions of a feature told from the perspective of the user. User stories can help you understand and prioritize the needs and wants of your users, and guide the development of the program. Identifying the users of the program and What are their goals and needs are can help understand what to include in user stories. 

How do I approach developing programs? What agile processes do I hope to incorporate into my future development work? 

I first Identify the problem or need that the program will solve when developing programs and write a list of requirements for the program to be finished. I then design a plan for how to design the program, and visualize the program's functionality and layout. For my future development work I hope to incorporate agile-scrum frameworks. I enjoy the collaborative process that this framework enforces and the process can be helpful when planning and implementing programs. 

What does it mean to be a good team member in software development?

Being a good team member involves effective communication and collaboration. Contributing your own skills and expertise while also being open to the ideas and contributions of others is an important trait of a good team member in software development. Being responsible and taking ownership of your work as well as being adaptable to any changes is also important to be a good team member.